File exist condition
This condition will detect if the file exist, if the file exist is not necessary to do all  the process to obtain the HTML code from the page. If the file does not exist it will do all the process. In both cases it will call the function to obtain a more specific data, like time or date et.